requirements.txt
SQLAlchemy==1.4.36
decorator==5.1.1
httpx==0.23.0
pytest==7.1.2
allure-pytest==2.9.45
python-dotenv==0.20.0
pytest-httpx==0.21.0